pyproject.toml
[tool.poetry]
name = "pandasai-to-pandas"
version = "0.1.0"
description = ""
authors = [
    "gumayusi1111 <zhaojunxi222@gmail.com>"
]
packages = [{include = "pandasai_to_pandas", from = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pandas = "1.5.3"
pandasai = "2.4.2"
python-dotenv = ">=1.1.0,<2.0.0"
numpy = "1.23.5"
flask = "^3.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
pytest = "^8.3.5"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"